+++ lib/Gortobot_library/gb_fix.h
@@ -6,13 +6,13 @@
 struct GbFix {
   float latitude;
   float longitude;
-  unsigned int year;
-  byte month;
-  byte day;
-  byte hour;
-  byte minute;
-  byte second;
-  byte satellites;
+  uint16_t year;
+  uint8_t month;
+  uint8_t day;
+  uint8_t hour;
+  uint8_t minute;
+  uint8_t second;
+  uint8_t satellites;
 };
 
 #endif
