+++ lib/Gortobot_library/gb_gps.h
@@ -8,8 +8,8 @@
 class GbGps {
 
 private:
-    byte _gps_power_pin1;
-    byte _gps_power_pin2;
+    uint8_t _gps_power_pin1;
+    uint8_t _gps_power_pin2;
     HardwareSerial* _gps_port;
     void GpsOn();
     void GpsOff();
@@ -18,7 +18,7 @@ private:
 
 public:
     //NMEAGPS _gps;
-    GbGps(byte power_pin1, byte power_pin2, HardwareSerial &port, unsigned long baud);
+    GbGps(uint8_t power_pin1, uint8_t power_pin2, HardwareSerial &port, uint32_t baud);
     GbFix GetFix(char fix_type);
 };
