+++ src/communications/gb_sailing_orders.h
@@ -7,11 +7,11 @@ struct GbSailingOrders {
     // TODO enum these
     // Sail modes: r=real, t=testing, s=direct set, p=pulse, f=fake
     char sailMode;
-    unsigned long loggingInterval;
-    int orderedSailPositionA;
-    int orderedTackTimeA;
-    int orderedSailPositionB;
-    int orderedTackTimeB;
+    uint32_t loggingInterval;
+    uint16_t orderedSailPositionA;
+    uint16_t orderedTackTimeA;
+    uint16_t orderedSailPositionB;
+    uint16_t orderedTackTimeB;
 };
 
 #endif
